backend/pyproject.toml
[project]
name = "echo-swarm-backend"
version = "0.1.0"
description = "ECHO-SWARM backend: orchestration layer for crisis simulation"
requires-python = ">=3.11"

dependencies = [
    # API
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "websockets>=12.0",
    "httpx>=0.27",
    # Config
    "pydantic-settings>=2.2",
    # Redis
    "redis>=5.0",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "geoalchemy2>=0.15",
    # Auth
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
